requirements.txt
# AWS and Bedrock dependencies
boto3>=1.28.0
anthropic-bedrock>=0.1.0

# Google API dependencies
google-api-python-client>=2.80.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Utility dependencies
python-dotenv>=1.0.0
requests>=2.28.2
pyyaml>=6.0
logging>=0.5.1.2

# For any future data processing
pandas>=1.5.3
